+++ DataAccessLayer/Repository/IProjectsDataAccess.cs
@@ -6,8 +6,6 @@ namespace DataAccessLayer
 {
     public interface IProjectsDataAccess
     {      
-
-        IEnumerable<Projects> GetProjects(long personID, bool includePersonal);
-        
+        IEnumerable<Projects> GetProjects(uint personID, bool includePersonal);
     }
 }
